--- src/components/ViewHistory/ViewHistoryItem.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-import { Typography } from "neetoui";
-
-const ViewHistoryItem = React.forwardRef(
-  ({ id, selectedMovieId, title }, ref) => (
-    <div
-      ref={ref}
-      className={`m-2 rounded-lg text-center shadow-md ${
-        id === selectedMovieId ? "bg-blue-700 text-white" : "bg-blue-200"
-      }`}
-    >
-      <Typography className="truncate p-2 font-medium">{title}</Typography>
-    </div>
-  )
-);
-
-ViewHistoryItem.displayName = "ViewHistoryItem";
-
-export default ViewHistoryItem;
